+++ packages/data-curation/src/data_emulation/schemas/enhancement_type.py
