+++ packages/data-curation/src/data_emulation/tsdae.py
